+++ src/Oro/Bundle/UserBundle/Tests/Unit/Form/Type/AclRoleTypeTest.php
@@ -2,7 +2,7 @@
 
 namespace Oro\Bundle\UserBundle\Tests\Unit\Form\Type;
 
-use Oro\Bundle\UserBundle\Form\Type\AclRoleType;
+use Pim\Bundle\UserBundle\Form\Type\AclRoleType;
 
 class AclRoleTypeTest extends \PHPUnit_Framework_TestCase
 {
