+++ src/Pim/Bundle/UserBundle/Form/Subscriber/PatchSubscriber.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Oro\Bundle\UserBundle\Form\EventListener;
+namespace Pim\Bundle\UserBundle\Form\Subscriber;
 
 use Doctrine\Common\Collections\Collection;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
